Create BobbysHobbies
In this project, we have created a table called "Persons"  which contains the name and age of each person, alongside a table called "Hobbies" which also lists each person's hobbies. We then joined the tables and queried them to show each person's name and hobby next to each other, and lastly we narrowed it down further to only show Bobby McBobbyFace's favorite hobby.