Create BandsInfo
In this project, I created three tables. One table lists the name of a band and one of their popular albums. The second table lists the popular albums and the year they came out. And the third table lists the single from each of those albums and the length of the song. I then joined the albums to make a new table that would show the band name, the album title and the single from each album.